aula 03

+++ server/src/routes/memories.ts
@@ -4,8 +4,15 @@ import { z } from 'zod'
 import { FastifyInstance } from 'fastify'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler' ,async (request) => {
+     await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+        where: {
+          userId: request.user.sub,
+        },
         orderBy: {
             createdAt: 'asc',
         },
@@ -20,7 +27,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     // const { id } = request.params
 
     const paramsSchema = z.object({
@@ -35,46 +42,60 @@ export async function memoriesRoutes(app: FastifyInstance) {
         },
     })
 
+    if (!memory.isPublic && memory.userId !== request.user.sub ) {
+      return reply.status(401).send()
+    }
+
     return memory
   })
 
   app.post('/memories', async (request) => {
-    const BodySchema = z.object({
+    const bodySchema = z.object({
         content: z.string(),
         coverUrl: z.string(),
         isPublic: z.coerce.boolean().default(false),
     })
 
-    const { content, coverUrl, isPublic } = BodySchema.parse(request.body)
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
     const memory = await prisma.memory.create({
       data: {
         content,
         coverUrl,
         isPublic,
-        userId: '0711e1d7-4590-41df-a1e1-967941daf3ba',
-      }
+        userId: request.user.sub,
+      },
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
   
     const { id } = paramsSchema.parse(request.params)
     
-    const BodySchema = z.object({
+    const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
     })
 
-    const { content, coverUrl, isPublic } = BodySchema.parse(request.body)
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub ){
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -88,12 +109,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
   
   const { id } = paramsSchema.parse(request.params)
+
+  const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub ){
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
